Remove usage of AlertsSection and replace it with Alerts component
