+++ src/components/clusters/NewClusterPage.tsx
@@ -16,7 +16,7 @@ import {
 } from '@patternfly/react-core';
 import { ToolbarButton } from '../ui/Toolbar';
 import ClusterToolbar from './ClusterToolbar';
-import AlertsSection from '../ui/AlertsSection';
+import Alerts from '../ui/Alerts';
 import { handleApiError, getErrorMessage } from '../../api/utils';
 import { AlertsContext, AlertsContextProvider } from '../AlertsContextProvider';
 import ClusterBreadcrumbs from './ClusterBreadcrumbs';
@@ -121,6 +121,7 @@ const NewClusterForm: React.FC<NewClusterFormProps> = ({ pullSecret = '' }) => {
                 </Stack>
               </GridItem>
             </Grid>
+            <Alerts />
             <ClusterToolbar>
               <ToolbarButton
                 name="save"
@@ -156,7 +157,6 @@ const NewCluster: React.FC = () => {
       </PageSection>
       <PageSection variant={PageSectionVariants.light} isFilled>
         {pullSecret === undefined ? <LoadingState /> : <NewClusterForm pullSecret={pullSecret} />}
-        <AlertsSection />
       </PageSection>
     </>
   );
