--- src/components/ui/AlertsSection/AlertsSection.css
@@ -1,6 +1,0 @@
-.pf-c-alert-group.alerts-section {
-  position: absolute;
-  padding: var(--pf-global--spacer--lg);
-  bottom: 0;
-  width: 100%;
-}
